[Refactor] move _run_stage logic into stage object
 - extract class Stage and StageFile into stage.py
